Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

This is a full restore to stock config, which I'm sure isn't necessary.

Binding the service to vsnet0, for example, breaks AGPS on Wifi, and requiring TLS is useless since we're not bundling the required certificates at /system/etc/cert.

You're changing a lot of options here, and I suggest you try to narrow it down to those effectively necessary.

Patch-set: 1