Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

This device has a dedicated partition for internal storage, unlike the Nexuses and most newer devices, which have a larger data partition and uses /data/media as a virtual sdcard. From what I can see in this patch, by moving the sdcard storage to /data/media, the internal sdcard partition is completely disabled, which I don't think is a good idea. It just leaves a lot of space unused, and wastes space in /data.

Patch-set: 2